[MOVE/#14] 네이밍 수정 및 폴더 이동
