--- app/src/main/java/org/sopt/cds29cm/data/CategoryFilter.kt
@@ -1,5 +1,0 @@
-package org.sopt.cds29cm.data
-
-data class CategoryFilter(
-    val filterName: String,
-)
