+++ app/src/main/java/org/sopt/cds29cm/data/dataclass/HatCategoryFilter.kt
@@ -1,0 +1,5 @@
+package org.sopt.cds29cm.data.dataclass
+
+data class HatCategoryFilter(
+    val filterName: String,
+)
