+++ app/src/main/java/org/sopt/cds29cm/data/dto/request/RequestHeartDTO.kt
@@ -1,4 +1,4 @@
-package org.sopt.cds29cm.data
+package org.sopt.cds29cm.data.dto.request
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
