+++ app/src/main/java/org/sopt/cds29cm/presentation/category/CategoryVerticalLeftAdapter.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import org.sopt.cds29cm.data.CategoryVertical
+import org.sopt.cds29cm.data.dataclass.CategoryVertical
 import org.sopt.cds29cm.databinding.ItemCategoryVerticalLeftBinding
 
 class CategoryVerticalLeftAdapter(context: Context) :
