+++ app/src/main/java/org/sopt/cds29cm/presentation/category/CategoryVerticalLeftViewHolder.kt
@@ -1,7 +1,8 @@
 package org.sopt.cds29cm.presentation.category
 
 import androidx.recyclerview.widget.RecyclerView
-import org.sopt.cds29cm.data.CategoryVertical
+import org.sopt.cds29cm.R
+import org.sopt.cds29cm.data.dataclass.CategoryVertical
 import org.sopt.cds29cm.databinding.ItemCategoryVerticalLeftBinding
 
 class CategoryVerticalLeftViewHolder(private val binding: ItemCategoryVerticalLeftBinding) :
