+++ app/src/main/java/org/sopt/cds29cm/presentation/hatCategory/HatCategoryBestItemViewHolder.kt
@@ -1,12 +1,12 @@
 package org.sopt.cds29cm.presentation.hatCategory
 
 import androidx.recyclerview.widget.RecyclerView
-import org.sopt.cds29cm.data.BestItem
+import org.sopt.cds29cm.data.dataclass.HatBestItem
 import org.sopt.cds29cm.databinding.ItemHatCategoryBestHatBinding
 
 class HatCategoryBestItemViewHolder(private val binding: ItemHatCategoryBestHatBinding) :
     RecyclerView.ViewHolder(binding.root) {
-    fun onBind(bestItemDataList: BestItem) {
+    fun onBind(bestItemDataList: HatBestItem) {
         with(binding) {
             tvHatCategoryBestItemDiscount.text = bestItemDataList.itemDiscount
             tvHatCategoryBestItemName.text = bestItemDataList.itemName
